adding example of using manageritm
launches:
1. manageritm server in a docker container
2. selenium grid and web browsers docker containers
3. example script in docker container

example script configures a web browser to talk to the manageritm
server, and launches the web browser. while browser is running, user can
interact with it through a vnc viewer. when the user is done, they send
a CTRL-C to the example script. the example script closes the web
browser and tells manageritm to shut down the proxy.