feat: user can specify flags to proxy, and specify command
When using the /client/proxy endpoint, users can send additional command
line options to the proxy command. In the data sent with the request,
provide a key named `additional_flags` and a value that is a list of
flags to be appended to the default command.

Example:

```
{
   ...
   'additional_flags': ['--opt1','value1','--opt2','value2']
}
```

When using the /client/command endpoint, users can set the command to
run. In the data sent with the request, provide a key named `command`
and a value that that is a list version of the command to run.

Example:

```
{
   ...
   'command': ['sleep','200000000000']
}
```

Added test cases for the new features.

Updating Makefile to use `pyenv virtualenv-delete` to delete the virtual
environment instead of `pyenv uninstall`. Also, I think the hooks are
properly being created in pyenv now, so we don't need to install black,
pytest, and pdbpp outside of poetry. We can run `poetry install` to get
all of the pieces installed.